+++ src/randomstate.rs
@@ -1,5 +1,5 @@
 #[derive(Copy, Clone)]
 #[repr(C)]
 pub struct RandomState {
-    pub data: [u64; 4],
+    pub data: [usize; 4],
 }
